Revert "changed molbe.BE.localize to be properly defined as method"
This reverts commit f917b9f5dd15c4dde0af72d4f943988ce89c2366.
