+++ src/quemb/molbe/mbe.py
@@ -1,30 +1,14 @@
 # Author(s): Oinam Romesh Meitei
 
-import functools
 import os
 import pickle
-import sys
 
 import h5py
 import numpy
-from numpy.linalg import eigh, multi_dot, svd
 from pyscf import ao2mo
 
 from quemb.molbe.be_parallel import be_func_parallel
 from quemb.molbe.eri_onthefly import integral_direct_DF
-from quemb.molbe.external.lo_helper import (
-    get_aoind_by_atom,
-    reorder_by_atom_,
-)
-from quemb.molbe.helper import ncore_, unused
-from quemb.molbe.lo import (
-    get_iao,
-    get_loc,
-    get_pao,
-    get_pao_native,
-    get_xovlp,
-    remove_core_mo,
-)
 from quemb.molbe.pfrag import Frags
 from quemb.molbe.solver import be_func
 from quemb.shared import be_var
@@ -320,273 +304,8 @@ class BE:
     #  cannot be moved to head of file.
     from quemb.molbe._opt import optimize  # noqa: PLC0415
     from quemb.molbe.external.optqn import get_be_error_jacobian  # noqa: PLC0415
+    from quemb.molbe.lo import localize  # noqa: PLC0415
     from quemb.molbe.rdm import compute_energy_full, rdm1_fullbasis  # noqa: PLC0415
-
-    def localize(
-        self,
-        lo_method,
-        mol=None,
-        valence_basis="sto-3g",
-        hstack=False,
-        pop_method=None,
-        init_guess=None,
-        valence_only=False,
-        nosave=False,
-    ):
-        """Molecular orbital localization
-
-        Performs molecular orbital localization computations. For large basis,
-        IAO is recommended augmented with PAO orbitals.
-
-        Parameters
-        ----------
-        lo_method : str
-        Localization method in quantum chemistry. 'lowdin', 'boys', and 'iao'
-        are supported.
-        mol : pyscf.gto.Molecule
-        pyscf.gto.Molecule object.
-        valence_basis: str
-        Name of minimal basis set for IAO scheme. 'sto-3g' suffice for most cases.
-        valence_only: bool
-        If this option is set to True, all calculation will be performed in the valence
-        basis in the IAO partitioning.
-        This is an experimental feature.
-        """
-        if lo_method == "lowdin":
-            es_, vs_ = eigh(self.S)
-            edx = es_ > 1.0e-15
-            self.W = numpy.dot(vs_[:, edx] / numpy.sqrt(es_[edx]), vs_[:, edx].T)
-            if self.frozen_core:
-                if self.unrestricted:
-                    P_core = [
-                        numpy.eye(self.W.shape[0]) - numpy.dot(self.P_core[s], self.S)
-                        for s in [0, 1]
-                    ]
-                    C_ = numpy.dot(P_core, self.W)
-                    Cpop = [
-                        functools.reduce(numpy.dot, (C_[s].T, self.S, C_[s]))
-                        for s in [0, 1]
-                    ]
-                    Cpop = [numpy.diag(Cpop[s]) for s in [0, 1]]
-                    no_core_idx = [numpy.where(Cpop[s] > 0.7)[0] for s in [0, 1]]
-                    C_ = [C_[s][:, no_core_idx[s]] for s in [0, 1]]
-                    S_ = [
-                        functools.reduce(numpy.dot, (C_[s].T, self.S, C_[s]))
-                        for s in [0, 1]
-                    ]
-                    W_ = []
-                    for s in [0, 1]:
-                        es_, vs_ = eigh(S_[s])
-                        s_ = numpy.sqrt(es_)
-                        s_ = numpy.diag(1.0 / s_)
-                        W_.append(functools.reduce(numpy.dot, (vs_, s_, vs_.T)))
-                    self.W = [numpy.dot(C_[s], W_[s]) for s in [0, 1]]
-                else:
-                    P_core = numpy.eye(self.W.shape[0]) - numpy.dot(self.P_core, self.S)
-                    C_ = numpy.dot(P_core, self.W)
-                    # NOTE: PYSCF has basis in 1s2s3s2p2p2p3p3p3p format
-                    # fix no_core_idx - use population for now
-                    Cpop = functools.reduce(numpy.dot, (C_.T, self.S, C_))
-                    Cpop = numpy.diag(Cpop)
-                    no_core_idx = numpy.where(Cpop > 0.7)[0]
-                    C_ = C_[:, no_core_idx]
-                    S_ = functools.reduce(numpy.dot, (C_.T, self.S, C_))
-                    es_, vs_ = eigh(S_)
-                    s_ = numpy.sqrt(es_)
-                    s_ = numpy.diag(1.0 / s_)
-                    W_ = functools.reduce(numpy.dot, (vs_, s_, vs_.T))
-                    self.W = numpy.dot(C_, W_)
-
-            if self.unrestricted:
-                if self.frozen_core:
-                    self.lmo_coeff_a = functools.reduce(
-                        numpy.dot, (self.W[0].T, self.S, self.C_a[:, self.ncore :])
-                    )
-                    self.lmo_coeff_b = functools.reduce(
-                        numpy.dot, (self.W[1].T, self.S, self.C_b[:, self.ncore :])
-                    )
-                else:
-                    self.lmo_coeff_a = functools.reduce(
-                        numpy.dot, (self.W.T, self.S, self.C_a)
-                    )
-                    self.lmo_coeff_b = functools.reduce(
-                        numpy.dot, (self.W.T, self.S, self.C_b)
-                    )
-            else:
-                if self.frozen_core:
-                    self.lmo_coeff = functools.reduce(
-                        numpy.dot, (self.W.T, self.S, self.C[:, self.ncore :])
-                    )
-                else:
-                    self.lmo_coeff = multi_dot((self.W.T, self.S, self.C))
-
-        elif lo_method in ["pipek-mezey", "pipek", "PM"]:
-            es_, vs_ = eigh(self.S)
-            edx = es_ > 1.0e-15
-            self.W = numpy.dot(vs_[:, edx] / numpy.sqrt(es_[edx]), vs_[:, edx].T)
-
-            es_, vs_ = eigh(self.S)
-            edx = es_ > 1.0e-15
-            W_ = numpy.dot(vs_[:, edx] / numpy.sqrt(es_[edx]), vs_[:, edx].T)
-            if self.frozen_core:
-                P_core = numpy.eye(W_.shape[0]) - numpy.dot(self.P_core, self.S)
-                C_ = numpy.dot(P_core, W_)
-                Cpop = functools.reduce(numpy.dot, (C_.T, self.S, C_))
-                Cpop = numpy.diag(Cpop)
-                no_core_idx = numpy.where(Cpop > 0.55)[0]
-                C_ = C_[:, no_core_idx]
-                S_ = functools.reduce(numpy.dot, (C_.T, self.S, C_))
-                es_, vs_ = eigh(S_)
-                s_ = numpy.sqrt(es_)
-                s_ = numpy.diag(1.0 / s_)
-                W_ = functools.reduce(numpy.dot, (vs_, s_, vs_.T))
-                W_ = numpy.dot(C_, W_)
-
-            self.W = get_loc(
-                self.mol, W_, "PM", pop_method=pop_method, init_guess=init_guess
-            )
-
-            if not self.frozen_core:
-                self.lmo_coeff = self.W.T @ self.S @ self.C
-            else:
-                self.lmo_coeff = self.W.T @ self.S @ self.C[:, self.ncore :]
-
-        elif lo_method == "iao":
-            loc_type = "SO"
-            val_basis = "sto-3g"
-
-            # Occupied mo_coeff (with core)
-            Co = self.C[:, : self.Nocc]
-            # Get necessary overlaps, second arg is IAO basis
-            S12, S2 = get_xovlp(self.mol, basis=val_basis)
-            # Use these to get IAOs
-            Ciao = get_iao(Co, S12, self.S, S2=S2)
-
-            if not valence_only:
-                # Now get PAOs
-                if loc_type.upper() != "SO":
-                    Cpao = get_pao(Ciao, self.S, S12, S2, self.mol)
-                elif loc_type.upper() == "SO":
-                    Cpao = get_pao_native(
-                        Ciao, self.S, self.mol, valence_basis=val_basis
-                    )
-
-            # rearrange by atom
-            aoind_by_atom = get_aoind_by_atom(self.mol)
-            Ciao, iaoind_by_atom = reorder_by_atom_(Ciao, aoind_by_atom, self.S)
-
-            if not valence_only:
-                Cpao, paoind_by_atom = reorder_by_atom_(Cpao, aoind_by_atom, self.S)
-
-            if self.frozen_core:
-                # Remove core MOs
-                Cc = self.C[:, : self.ncore]  # Assumes core are first
-                Ciao = remove_core_mo(Ciao, Cc, self.S)
-
-            # Localize orbitals beyond symm orth
-            if loc_type.upper() != "SO":
-                Ciao = get_loc(self.mol, Ciao, loc_type)
-                if not valence_only:
-                    Cpao = get_loc(self.mol, Cpao, loc_type)
-
-            shift = 0
-            ncore = 0
-            if not valence_only:
-                Wstack = numpy.zeros(
-                    (Ciao.shape[0], Ciao.shape[1] + Cpao.shape[1])
-                )  # -self.ncore))
-            else:
-                Wstack = numpy.zeros((Ciao.shape[0], Ciao.shape[1]))
-
-            if self.frozen_core:
-                for ix in range(self.mol.natm):
-                    nc = ncore_(self.mol.atom_charge(ix))
-                    ncore += nc
-                    niao = len(iaoind_by_atom[ix])
-                    iaoind_ix = [i_ - ncore for i_ in iaoind_by_atom[ix][nc:]]
-                    Wstack[:, shift : shift + niao - nc] = Ciao[:, iaoind_ix]
-                    shift += niao - nc
-                    if not valence_only:
-                        npao = len(paoind_by_atom[ix])
-                        Wstack[:, shift : shift + npao] = Cpao[:, paoind_by_atom[ix]]
-                        shift += npao
-            else:
-                if not hstack:
-                    for ix in range(self.mol.natm):
-                        niao = len(iaoind_by_atom[ix])
-                        Wstack[:, shift : shift + niao] = Ciao[:, iaoind_by_atom[ix]]
-                        shift += niao
-                        if not valence_only:
-                            npao = len(paoind_by_atom[ix])
-                            Wstack[:, shift : shift + npao] = Cpao[
-                                :, paoind_by_atom[ix]
-                            ]
-                            shift += npao
-                else:
-                    Wstack = numpy.hstack((Ciao, Cpao))
-            if not nosave:
-                self.W = Wstack
-                assert numpy.allclose(
-                    self.W.T @ self.S @ self.W, numpy.eye(self.W.shape[1])
-                )
-            else:
-                assert numpy.allclose(
-                    Wstack.T @ self.S @ Wstack, numpy.eye(Wstack.shape[1])
-                )
-                return Wstack
-            nmo = self.C.shape[1] - self.ncore
-            nlo = self.W.shape[1]
-
-            if not valence_only:
-                if nmo > nlo:
-                    Co_nocore = self.C[:, self.ncore : self.Nocc]
-                    Cv = self.C[:, self.Nocc :]
-                    # Ensure that the LOs span the occupied space
-                    assert numpy.allclose(
-                        numpy.sum((self.W.T @ self.S @ Co_nocore) ** 2.0),
-                        self.Nocc - self.ncore,
-                    )
-                    # Find virtual orbitals that lie in the span of LOs
-                    u, l, vt = svd(self.W.T @ self.S @ Cv, full_matrices=False)
-                    unused(u)
-                    nvlo = nlo - self.Nocc - self.ncore
-                    assert numpy.allclose(numpy.sum(l[:nvlo]), nvlo)
-                    C_ = numpy.hstack([Co_nocore, Cv @ vt[:nvlo].T])
-                    self.lmo_coeff = self.W.T @ self.S @ C_
-                else:
-                    self.lmo_coeff = self.W.T @ self.S @ self.C[:, self.ncore :]
-            else:
-                self.lmo_coeff = self.W.T @ self.S @ self.C[:, self.ncore :]
-
-        elif lo_method == "boys":
-            es_, vs_ = eigh(self.S)
-            edx = es_ > 1.0e-15
-            W_ = numpy.dot(vs_[:, edx] / numpy.sqrt(es_[edx]), vs_[:, edx].T)
-            if self.frozen_core:
-                P_core = numpy.eye(W_.shape[0]) - numpy.dot(self.P_core, self.S)
-                C_ = numpy.dot(P_core, W_)
-                Cpop = functools.reduce(numpy.dot, (C_.T, self.S, C_))
-                Cpop = numpy.diag(Cpop)
-                no_core_idx = numpy.where(Cpop > 0.55)[0]
-                C_ = C_[:, no_core_idx]
-                S_ = functools.reduce(numpy.dot, (C_.T, self.S, C_))
-                es_, vs_ = eigh(S_)
-                s_ = numpy.sqrt(es_)
-                s_ = numpy.diag(1.0 / s_)
-                W_ = functools.reduce(numpy.dot, (vs_, s_, vs_.T))
-                W_ = numpy.dot(C_, W_)
-
-            self.W = get_loc(self.mol, W_, "BOYS")
-
-            if not self.frozen_core:
-                self.lmo_coeff = self.W.T @ self.S @ self.C
-            else:
-                self.lmo_coeff = self.W.T @ self.S @ self.C[:, self.ncore :]
-
-        else:
-            print("lo_method = ", lo_method, " not implemented!", flush=True)
-            print("exiting", flush=True)
-            sys.exit()
 
     def print_ini(self):
         """
